+++ src/forze_fastapi/middlewares/__init__.py
@@ -1,0 +1,5 @@
+from .logging import LoggingMiddleware
+
+# ----------------------- #
+
+__all__ = ["LoggingMiddleware"]